# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrepMind AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExamRequest {
  examName: string;
}

export type PracticeRequestDifficulty =
  (typeof PracticeRequestDifficulty)[keyof typeof PracticeRequestDifficulty];

export const PracticeRequestDifficulty = {
  Easy: "Easy",
  Medium: "Medium",
  Hard: "Hard",
} as const;

export interface PracticeRequest {
  examName: string;
  topic: string;
  difficulty: PracticeRequestDifficulty;
  count?: number;
}

export interface MCQQuestion {
  id: string;
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
  trick?: string;
  topic: string;
  difficulty: string;
}

export interface PracticeResponse {
  questions: MCQQuestion[];
}

export interface TeachRequest {
  examName: string;
  topic: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
  systemContext?: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}