# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PrepMind AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prepmind/overview:
    post:
      operationId: getExamOverview
      tags: [prepmind]
      summary: Get exam overview, syllabus, pattern, strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamRequest"
      responses:
        "200":
          description: Exam overview streamed as SSE
          content:
            text/event-stream: {}
  /prepmind/practice:
    post:
      operationId: generatePracticeQuestions
      tags: [prepmind]
      summary: Generate MCQ practice questions for a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeRequest"
      responses:
        "200":
          description: Generated MCQ questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeResponse"
  /prepmind/teach:
    post:
      operationId: teachTopic
      tags: [prepmind]
      summary: Teach a specific topic with explanations, formulas, shortcuts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeachRequest"
      responses:
        "200":
          description: Teaching content streamed as SSE
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    ExamRequest:
      type: object
      properties:
        examName:
          type: string
      required:
        - examName
    PracticeRequest:
      type: object
      properties:
        examName:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        count:
          type: integer
      required:
        - examName
        - topic
        - difficulty
    PracticeResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/MCQQuestion"
      required:
        - questions
    MCQQuestion:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        explanation:
          type: string
        trick:
          type: string
        topic:
          type: string
        difficulty:
          type: string
      required:
        - id
        - question
        - options
        - correctIndex
        - explanation
        - topic
        - difficulty
    TeachRequest:
      type: object
      properties:
        examName:
          type: string
        topic:
          type: string
      required:
        - examName
        - topic
...