# lib/api-zod/src/generated/types/mCQQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrepMind AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MCQQuestion {
  id: string;
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
  trick?: string;
  topic: string;
  difficulty: string;
}