# lib/api-zod/src/generated/types/practiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrepMind AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeRequestDifficulty } from "./practiceRequestDifficulty";

export interface PracticeRequest {
  examName: string;
  topic: string;
  difficulty: PracticeRequestDifficulty;
  count?: number;
}