# lib/api-zod/src/generated/types/practiceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrepMind AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MCQQuestion } from "./mCQQuestion";

export interface PracticeResponse {
  questions: MCQQuestion[];
}